updated file structure, removed unnecessary code, created separate components for header, search and actions
